+++ chouette/src/main/java/fr/certu/chouette/service/validation/util/GroupOfLineProducer.java
@@ -66,7 +66,7 @@ class GroupOfLineProducer {
 		
 		// ObjectVersion optionnel
 		if (castorGroupOfLine.hasObjectVersion()) {
-			int castorObjectVersion = castorGroupOfLine.getObjectVersion();
+			int castorObjectVersion = (int)castorGroupOfLine.getObjectVersion();
 			if (castorObjectVersion < 0) {
 				if (castorGroupOfLine.getName() != null)
 					params = LoggingManager.getParams(""+castorObjectVersion, castorGroupOfLine.getName());
